refactor: #119 - 커스텀훅 추가 및 컴포넌트 이사
