+++ src/components/PostViewer/components/index.ts
@@ -1,0 +1,4 @@
+export { default as ButtonGroup } from './ButtonGroup';
+export { default as NonAuthModal } from './NonAuthModal';
+export { default as PostContents } from './PostContents';
+export { default as PostTitle } from './PostTitle';
