+++ src/components/PostViewer/hooks/index.ts
@@ -1,4 +1,5 @@
 export { useElementWidth } from './useElementWidth';
+export { useControlRouteByPost } from './useControlRouteByPost';
 export { useContentClassName } from './useContentClassName';
 export { usePostLike } from './usePostLike';
 export { useActionInfo } from './useActionInfo';
